Move fuzzing tests to acceptance
